requirements.txt
asgiref==3.9.1
attrs==25.3.0
cfgv==3.4.0
click==8.2.1
distlib==0.4.0
Django==5.2.6
django-cors-headers==4.8.0
django-filter==25.1
djangorestframework==3.16.1
PyJWT==2.8.0
drf-spectacular==0.26.5
filelock==3.19.1
h11==0.16.0
identify==2.6.14
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
nodeenv==1.9.1
platformdirs==4.4.0
pre_commit==4.3.0
psycopg==3.1.12
psycopg-binary==3.1.12
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
uvicorn==0.35.0
virtualenv==20.34.0
whitenoise==6.10.0
aiogram==3.2.0
redis==5.0.1
